requirements.txt
mcp[cli]>=1.6.0
smolagents[litellm,mcp]>=1.13.0
requests>=2.32.3
python-dotenv>=1.0.0
boto3>=1.38.43
botocore>=1.34.0
python-dateutil>=2.8.0
pytz>=2023.3
pyodbc>=5.1.0
PyYAML>=6.0
pytest>=8.0.0
pydantic-settings>=2.0.0
structlog>=24.1.0
